+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/estadoPedido/EstadoPedidoEntity.java
@@ -14,19 +14,19 @@ import javax.persistence.Table;
 public class EstadoPedidoEntity implements Serializable {
 
 	private static final long serialVersionUID = 1L;
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private int id;
+	private long id;
 
 	private String NombreEstado;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
