+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/formaPago/FormaPagoEntity.java
@@ -17,31 +17,36 @@ import javax.persistence.Table;
 public abstract class FormaPagoEntity implements Serializable {
 
 	private static final long serialVersionUID = -8356649232468048872L;
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private int id; 
-	
+	private long id;
+
 	protected String denominacion;
-	protected double montoDescuento; 
-	 
+	protected double montoDescuento;
+
 	public String getDenominacion() {
 		return denominacion;
 	}
+
 	public void setDenominacion(String denominacion) {
 		this.denominacion = denominacion;
 	}
+
 	public double getMontoDescuento() {
 		return montoDescuento;
 	}
+
 	public void setMontoDescuento(double montoDescuento) {
 		this.montoDescuento = montoDescuento;
 	}
-	public int getId() {
+
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+
+	public void setId(long id) {
 		this.id = id;
-	} 
+	}
 }
