+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/rol/RolEntity.java
@@ -12,14 +12,14 @@ import javax.persistence.Table;
 @Entity
 @Table(name = "rol")
 public class RolEntity implements Serializable {
-	
+
 	private static final long serialVersionUID = 5685412683798686451L;
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private int id;
-	private String nombreRol; 
+	private long id;
+	private String nombreRol;
 
 	public String getNombreRol() {
 		return nombreRol;
@@ -27,16 +27,14 @@ public class RolEntity implements Serializable {
 
 	public void setNombreRol(String nombreRol) {
 		this.nombreRol = nombreRol;
-	} 
+	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
-	
-	
 
 }
